+++ R/critVal-methods.R
@@ -1,0 +1,9 @@
+#' @rdname critVal
+setMethod("critVal", "ELT", function(object, ...) {
+  object@cv
+})
+
+#' @rdname critVal
+setMethod("critVal", "ELMT", function(object, ...) {
+  object@cv
+})